Refactoring: Move retrieving price in USD to hook
